don't remove important copyright information
Remove copyright "headers" such as "Copyright (c) <year> <author>" can
be a problem if the license text itself will be used later on, e.g. to
be used in a generated NOTICE file.

This is for the moment just a quick hack for a PoC, since while the
important information should now be kept, it will also obviously
decrease the accuracy of the classifier.